Corretto type Impegno

+++ src/main/java/it/cnr/si/missioni/util/proxy/json/object/ImpegnoGae.java
@@ -23,7 +23,7 @@ public class ImpegnoGae extends RestServiceBean  implements Serializable {
 	@JsonProperty("cdCdsOrigine")
 	private String cdCdsOrigine;
 	@JsonProperty("pgObbligazione")
-	private Integer pgObbligazione;
+	private Long pgObbligazione;
 	@JsonProperty("dsObbligazione")
 	private String dsObbligazione;
 	@JsonProperty("cdUoOrigine")
@@ -179,7 +179,7 @@ public class ImpegnoGae extends RestServiceBean  implements Serializable {
 	 * The pgObbligazione
 	 */
 	@JsonProperty("pgObbligazione")
-	public Integer getPgObbligazione() {
+	public Long getPgObbligazione() {
 		return pgObbligazione;
 	}
 
@@ -189,7 +189,7 @@ public class ImpegnoGae extends RestServiceBean  implements Serializable {
 	 * The pgObbligazione
 	 */
 	@JsonProperty("pgObbligazione")
-	public void setPgObbligazione(Integer pgObbligazione) {
+	public void setPgObbligazione(Long pgObbligazione) {
 		this.pgObbligazione = pgObbligazione;
 	}
